src/requirements.txt
requests==2.27.0
kafka-python==2.0.2
boto3==1.34.156
git+https://github.com/openskynetwork/opensky-api.git@master#subdirectory=python
